fully driven by state machine only
The loop() function still contained code to check the cardreader.

Moved that into the function read_loop() which is repeatedly
called as long as the stae is "ready".

Background: states have actually three callbacks: on_entry,
on_exit, and on_state. The former tow have been used already.
On_state is repeatedly called as long as teh machine is in the
respective state.