simplified finite state machine
There is only three states: ready, authenticated, and brewing.
Reneamed coffee_ready into brewing.
Dropped the states fail (never used) and user_unknown.

Most of the state machine is driven from ready_loop(). In
particular, the only purpose of checkForCard is to return a
token if card present or NO_CARD otherwise. Only then, is
authenticateToken() used to auth it and trigger AI:EVENT to
transit into authenticated state.

If a unkwon card is presented, it triggers a transition from
ready back to ready. The only purpose of this transition is
to be able to display a GUI message in teh transition's callback.

In general, sometimes it is more useful to employ callbacks on
transitions, rather than on entry and exit from states.