Added state machine to control gui.
It has three states:
* cleared
* delaying
* displaying

The state machine is designed with the future use of the coffee
machine's display in mind. The problem is, that the known command
for displaying messages blocks messages from the coffee machine.

Then the main purpose of the cleared state is to return control
over the display back to the coffee machine.

The state displaying takes control of the display to print a
message, but falles back to state cleared within a short time.

Delaying is just a conveninece to delay displaying non-urgent
messages for a few second. It is currently only used to avoid
overwritting messages when the coffee machine state machine
falls back into state ready.

A schematic of the gui state machine can be found in the code.